Split Dynamic attributes to its own interfaces
