+++ src/main/java/com/sigmundgranaas/forgero/toolhandler/DynamicDurability.java
@@ -1,0 +1,7 @@
+package com.sigmundgranaas.forgero.toolhandler;
+
+import net.minecraft.item.ItemStack;
+
+public interface DynamicDurability {
+    int getDurability(ItemStack stack);
+}
