+++ src/main/java/com/sigmundgranaas/forgero/toolhandler/DynamicEffectiveNess.java
@@ -1,0 +1,9 @@
+package com.sigmundgranaas.forgero.toolhandler;
+
+import net.minecraft.block.BlockState;
+
+public interface DynamicEffectiveNess {
+    default boolean isEffectiveOn(BlockState state) {
+        return false;
+    }
+}
