+++ src/main/java/com/sigmundgranaas/forgero/toolhandler/DynamicMiningLevel.java
@@ -1,0 +1,4 @@
+package com.sigmundgranaas.forgero.toolhandler;
+
+public interface DynamicMiningLevel {
+}
