+++ src/main/java/com/sigmundgranaas/forgero/toolhandler/DynamicMiningSpeed.java
@@ -1,0 +1,10 @@
+package com.sigmundgranaas.forgero.toolhandler;
+
+import net.minecraft.block.BlockState;
+import net.minecraft.item.ItemStack;
+
+public interface DynamicMiningSpeed {
+    default float getMiningSpeedMultiplier(BlockState state, ItemStack stack) {
+        return 1.0F;
+    }
+}
